funny 2am commit
why do I do this to myself ._.
- changed some pins to make them harder to plug i--uhh--to make them closer together!!11!
- add config for i2c pins
- people without a mode switch aren't screwed anymore
- reset code will actually reset now
- cured my programming ocd in a couple spots (not for long though)
- chatgpt told me to add break statements so I did
- tune mode was NOT integrated properly (and probably isn't now either)
- there was an error with the error (codes)
- oh my gosh I can finally go to bed

+++ src/modes/modes.h
@@ -11,7 +11,7 @@
  * Runs the specified mode of the system.
  * Additionally, if the system is not already set into the specified mode, it will run the mode starting code.
 */
-void mode(uint smode);
+void mode(uint8_t smode);
 
 /**
  * Declares whether or not the IMU data is safe to use.
